Build libinnobase.a for the innodb plugin, add facebook startup file
Summary: This makes libinnobase.a for the innodb plugin and links it to mysqld.
That avoids the need to run INSTALL commands to load the innodb shared object.
This builds innochecksum and installs innodb man pages when the plugin is used.
This also adds a facebook startup file.
This also patches mysql-test-run so that command line mysqld options,
such as --mysqld=--innodb, are used to update mysqld_variables and determine
which tests should not be skipped.
Pass --user=root when running mysqld from mysql-test-run to get the list of supported
options.

Reviewed By: ps

Test Plan: build an RPM