Updated innodb tests with results for innodb-plugin and information_schema.file_status.
Summary:
Some tests fail because of new features and defaults in the innodb
plugin.  This updates the various tests to the values returned by the
community-sources built with the innodb-plugin instead of builtin-innobase.
Additionally this includes the change to the results from the addition of the
information_schema.table_status table.

Reviewed By: mcallaghan

Test Plan:
This commit is a test plan.