Port Mark's FLIFO patches.
Summary: Add innodb_thread_lifo as additional thread scheduling policy
for innodb_thread_concurrency. The current policy is FIFO. This adds
LIFO. When enabled, both are used and I will call the result FLIFO.
This compensates for the long time for a thread to begin running
when waiting (alone) on a pthread condition variable and that a
pthread condition variable broadcast is done to wake it. Definitions
fifo-pending - the number of threads that have been selected to run
   via a pthread condition variable broadcast but have yet to begin
   running.
lifo-running - the number of threads that are currently running
   courtesy of the LIFO policy.
When a thread would not be allowed to enter InnoDB because of the
innodb_thread_concurrency limit, it is allowed to enter when
  lifo-running < fifo-pending.
...and
Note that this allows 2 * innodb_thread_concurrency threads to run
concurrently in some cases.
...and
Also note that when a thread is blocked on IO (usually a disk read),
it is still considered to be running within InnoDB. When a thread
is blocked waiting on a row lock, it 'exits' InnoDB.
...and
This makes performance
much better under high concurrency. For example, this is throughput
in transactions per second for sysbench oltp readonly at 64, 128, 256,
5120 and 1024 concurrent session. 5044 is unmodified MySQL 5.0.44
and 5084concur is MySQL 5.0.84 with the patch. Throughput is 2.45X
higher with the patch.
    664     597     534     439     260     5044
    575     590     611     613     637     5084concur
...and
This change also prevents innodb_thread_concurrency from begin
toggled between 0 and !0 when innodb_thread_lifo is set. I am not
sure if it should be toggled when innodb_thread_lifo is not set but
I will leave that for another day.
...and
This moves bookkeeping that was done by the thread scheduled to
run at the end of srv_conc_enter_innodb to the srv_conc_force_exit_innodb
to be done by the thread that schedules it. This removes the need
for the scheduled thread to lock srv_conc_mutex which is a very hot
mutex, so the scheduled thread should begin to run faster.

Reviewers:mcallaghan

Test Plan:mtr for now, need to do perf test later ot make sure it's working