Port v5.0 "Add my.cnf variable innodb_flush_neighbors_on_checkpoint"
Summary: Add my.cnf variable innodb_flush_neighbors_on_checkpoint. This
is TRUE by default to match default InnoDB behavior. When FALSE, dirty
neighbor pages in an extent are not written to disk during a checkpoint
when the min LSN of the oldest dirty page must be advanced. This can
reduce the rate at which page writes are done.

Reviewers:mcallaghan

Test Plan:mtr