Re-enable two tests and update for plugin.
Summary:The innodb-autoinc-44030 and innodb-index tests were disabled
because of some bugs that are now fixed in our branch.  I'm re-enabling
the tests.  I had to make a few small changes for differences in the
plugin.  For 44030 I have to reset the file format variable which
changes during server reboot.  For innodb-index it picks a different
(but equally valid) uniqueness constraint violation to complain about.
I've seen this change in test results previously and in the other case
the plugin result matched all other storage engines while built-in
innodb was the outlier, so I feel this change is correct.

Reviewers:mcallaghan

Test Plan:mtr