Port: Provide fast and non-blocking alternative to FLUSH TABLES WITH READ LOCK for hot backup
Summary:Provide fast and non-blocking alternative to FLUSH TABLES WITH READ LOCK for hot backup.
See mysqlha.blogspot.com/2009/10/be-careful-with-flush-tables-with-read.html for context.
This adds START TRANSACTION WITH CONSISTENT INNODB SNAPSHOT that starts an Innodb transaction
and returns the master binlog filename and offset. When this is used for backups, replication
should be restarted at those return values. The new START command returns 1 row with 2 columns
(File, Position) on success.

Reviewers:mcallaghan

Test Plan:mtr