Do not free table_stats hash when clearing it, allowing less locking.
Summary:
FLUSH STATISTICS used to empty the table_status hash of all
entries.  This caused additional complexity and locking to prevent a
thread from attempting to update the statistics while it was being
deleted.  This changes FLUSH STATISTICS to only reset the table_status
values to 0, not to free any memory.  This allows minimal locking (only
for the duration of the hash lookup and insertion, which is already
cached by the handler).  The actual updates (through the cached pointer)
are done atomically to avoid inaccuracies.

Reviewed By: mcallaghan

Test Plan:
mtr