Undo a spurious change to information_schema test results.
Summary:
I thought this was necessary because of the additional i_s table
but I think it was only different because I was running an unusual test
sequence and one of the tests doesn't properly clean up.  I'm restoring
it to the original value from the community sources and it runs cleanly.

Reviewed By: mcallaghan

Test Plan: