Group commit on binlog
Summary:
Add the possibility of group committing on the binlog. A new parameter,
binlog_sync_timeout_usecs, controls how long a thread will wait to fsync
to the binlog. After waiting that long, a thread will fsync, or, if a
previous thread fsyncs, the thread will continue.

Test Plan:
mtr
more tests to be written

Reviewed By: mcallaghan