Get and fix updated patch for mysql bug 51920 -- innodb ignores kill
Summary:
This applies and then fixes the updated patch for bug 51920. The
existing problem in the patch is that it returns -1 from
convert_error_code_to_mysql
and when -1 is returned, spam is written to the db error log.

I also changed the test to check that the returned error is "interrupted".

Test Plan:
mtr

Reviewed By: ryandmack