Restore some innodb-plugin changes to innodb tests lost in 5.1.47 merge
Summary:
5.1.47 moved innodb MTR tests to the suite/innodb directory. I
lost some local changes to make the tests pass against the plugin as
part of the merge.  I also checked in a merge conflict on the result
file that is now removed (because git intelligently finds moved files,
it's possible to have a merge conflict in a file marked as new, go
figure).

Test Plan:
mtr

Reviewed By: mcallaghan