Added the innodb_fast_checksums option.
Summary:
This added crc32 as an alternative to the normal innodb checksum
algorithm.  It is implemented operating on 8 bytes in parallel and uses
hardware crc32 instructions if available.  With hardware crc32 this is
about 10x faster than the old algorithm, and without it's 2x faster.
Once you run innodb with this enabled, old versions of innodb will
report corruption and must be run with checksumming disabled.

Test Plan:
Submitting now for early review.  MTR passed.  Will do
dump/reload testing and verify performance delta this afternoon.

Reviewed By: mcallaghan