Basis for per-user statistics
Summary:
More code will follow. This adds a struct to USER_CONN to track
statistics. check_user() is changed to get the USER_CONN struct sooner so
it can be updated on denied connection errors.  check_max_user_connections
is changed to not decrement as that is done in fix_user_conn.

Test Plan:
mtr

Reviewed By: ryandmack