Make innodb work for 8kb pages
Summary:
Without this change the allocation done from ha_create when
buf_pool_init calls it is attempted using buffer pool pages and that
function attempts to recursively lock the buffer pool mutex. It is
done from the buffer pool because len is not less than 1/2 of a page
in mem_heap_create_block. For 16kb pages, this allocation was less
than 1/2 of a page and ended up use MEM_HEAP_DYNAMIC.
...
Also change check for max log file size to work for any page size.

Test Plan:
mtr

Reviewed By: ryandmack