Reduce overhead from running buf_flush_free_margin
Summary:
buf_flush_free_margin is called to make sure there are
enough clean pages at the end of the LRU. It finds dirty pages and
issues write requests for them. This is done while holding the buffer
pool mutex and that is a big source of mutex contention especially on
an IO bound workload for an SMP server that can do a lot of IO. The
fix is:
* change callers to request a specific number of free pages. Most callers
only need one page from the free list.
* don't get the buffer pool mutex when the free list has enough pages. This
is done without getting the buffer pool mutex (...to avoid contention)
A feature requets is open for this -- http://bugs.mysql.com/bug.php?id=54815
With this change I was able to increase the max IOPs that InnoDB can do
on a 16-core server from 52k to 74k. This also dropped OS waits on the
buffer pool mutex by 50X when combined with earlier change to make
BUF_READ_AHEAD_AREA a constant.
...
A call to buf_flush_free_margin was put in srv_master_thread but is
currently commented out. I am not sure if it is needed. In case it is,
the counters below distinguish between calls to it from foreground and
background threads.
...
New counters in SHOW STATUS
...
Number of pages flushed by it:
    Innodb_buffer_pool_flushed_free_margin_foreground       0
    Innodb_buffer_pool_flushed_free_margin_background       0
Time spent in clals to buf_flush_free_margin
    Innodb_srv_free_margin_foreground_seconds       0.000006
    Innodb_srv_free_margin_background_seconds       0.000000
...
New data in SHOW INNODB STATUS
...
Time spent in calls to buf_flush_free_margin
-----------------
BACKGROUND THREAD
-----------------
Seconds in background IO: ... 0.00 free margin fg, 0.00 free margin bg, ...
...
In BUFFER POOL AND MEMORY, add counters for foreground and background calls to
buf_flush_free_margin
Write sources: fg free margin 0, bg free margin 0, ...

Test Plan:
mtr, benchmarks -- graphs to be published

Reviewed By: ryandmack