Merge changes from other FB init script, bump startup to 60s, shutdown to 10m
Summary:
We were having trouble bootstrapping a DB in 5.1 because of a
permissions issue on binlogs.  It turns out the fix (adding
--user=mysql) was already done in the copy of the init script in our
srpm repository but not this one that actually goes into the RPM.  I've
merged those changes into this file.  I've also bumped the startup
timeout to 60 seconds because bootstraping invokes mysql a few times and
each time now has an addition 2 second rdtsc calibration that was
causing the init script to fail on occasion.  Lastly I've bumped the
shutdown timeout to 10m to give innodb a better change to cleanly
shutdown when someone does an administrative reboot on a host.

Test Plan:
Install on a host, try bootstrapping a DB, shutting it down,
restarting it.

Reviewed By: mkonetchy