Prevent SELECT DATABASE() and @@version_comment from being sent when not required
Summary:
get_current_db() only needs to be used in the USE command when in interactive
or if one-database is specified.  Added check to bypass running that to avoid
unnecessary roundtrip to database.

Added a check for batch mode to prevent select @@version_comment from being
sent.  Output was being surpressed in put_info(), but the select was still being
run.

Test Plan:
I ran MySQL with the general log enabled to see all statements being sent to
the server.  The queries only showed up when appropriate.

./mysql --socket=/tmp/hfisk/mysql.sock -e "use test; select 1"

Modified client

100629 11:25:11	    5 Connect	hfisk@ as anonymous on
		    5 Init DB	test
		    5 Query	select 1
		    5 Quit

Original client

100629 11:25:20	    6 Connect	hfisk@ as anonymous on
		    6 Query	select @@version_comment limit 1
		    6 Query	SELECT DATABASE()
		    6 Init DB	test
		    6 Query	select 1
		    6 Quit

Reviewed By: dmituzas
Commenters: ps