Make BUF_READ_AHEAD_AREA a constant
Summary:
BUF_READ_AHEAD_AREA is frequently called by code that
holds the buffer pool mutex. It isn't cheap and the cost increases
the time for which the mutex is locked. Assuming there are enough
(a few thousand) pages in the buffer pool this can be a constant.
...
At high concurrency on a server doing a lot of IO this increased
QPS by 1.5X (40k to 60k).

Test Plan:
mtr, benchmarks (tests done, graphs in preparation)

Reviewed By: ryandmack