Migration support for broken slice8 implementaiton.
Summary:
Includes support for broken slice8 implementation on Nehalem to
allow migrations.  Support for the broken algorithm will be removed in
the next build and DBs should be dumped and reloaded to switch to the
correct implementation.  This will print a warning (once) if it
encounters a checksum computed by the broken function.

Test Plan:
Will test migrating a non-Nehalem broken DB to both a Nehalem
and non-Nehalem host.

Reviewed By: mcallaghan