[Kill server if there is an attempt to rollback after locks are released, fix an incorrect assert, fix couple of broken unit tests.]
Summary:
(1)
If innobase_release_locks_early is enabled, locks are released
at prepare time. If there is a need to rollback after that, it causes
problems because other transactions could have modified the rows that
need to be undone. Rolling back this xact could corrupt those rows/pages
and so kill the server.

Killing the server preserves data integrity at least in the case where
the log records of the other conflicting xacts have not yet been flushed
to disk. (If they have been flushed to disk, the problem may persist
during crash recovery as well.)

(2) My earlier change to print errors in log if current ticket does
not advance for a second is causing errors in force_binlog_order
and release_row_locks_early tests. So filter those errors.

(3) Fix an incorrect assert as mentioned in mysql issue
https://support.mysql.com/view.php?id=49670.

Test Plan:
I put back the bug mentioned in task 283270 and ran sysbench.
This time the server was killed and I did not see the row corruption
error such as [trying to access update undo rec field 32 in index "PRIMARY"
of table "sbtest"."sbtest" but index has only 6 fields].

Reviewed By: mcallaghan