Prevent mini-dogpiles from the per-table innodb statistics mutex
Summary:
There are two types of dogpiles for innodb statistics collection. The first
was previously fixed by us and occurs on database restart when multiple threads
collect index
stats for the same table. The 'force' parameter was added to
dict_update_statistics for that.

The new dogpile occurs when the stat_modification counter changes enough so
that stats must
be recollected. Alas, with a concurrent workload multiple sessions block in
dict_update_statistics_low
waiting for this to be done. It is better to let one session collect the stats
and all others
continue with their normal work. The 'wait' parameter was added to
dict_update_statistics to
request that behavior.

Finally, the stat_modification counter was incremented for every row update. It
only needs to
be incremented for updates to indexed columns.

Test Plan:
mtr, perf and stress tests to be done

Reviewed By: nponnekanti