Fix performance regression from change to make buf_flush_free_margin faster
Summary:
This fixes a performance regression from a previous change to make
buf_flush_free_margin
faster. The regression occurs on benchmarks that consume the free list quickly
-- IO bound,
buffer pool not large enough. The problem is that the number of pages moved to
the free list
used to be increased by BUF_FLUSH_EXTRA_MARGIN. Moving extra pages amortized
the cost of
getting the buffer pool mutex. This restores that and adds a call to
buf_flush_free_margin
to srv_master_thread. It also removes assignments in srv_master_thread that are
never read.
...
I have confirmed that this fixes the performance regression on IO-bound
sysbench oltp described at
www.facebook.com/notes/mysql-at-facebook/innodb-performance-55-versus-51-for-concurrent-writes-and-an-io-bound-server/434698460932
...
I am rerunning the original tests that justified this change.

Test Plan:
mtr, perf tests

Reviewed By: ryandmack