Add option to disable flushing of neighbor pages during LRU list flush
Summary:
buf_flush_free_margin is a huge source of contention on the buffer pool
mutex especially when many pages must be moved from the end of the LRU to the
free
list because of buffer pool pressure. Part of the problem is the duration for
which
the mutex is locked in buf_flush_try_neighbors. In this case the objective is
to
move N pages from the end of the LRU to the free list and those pages have
already
been found prior to calling buf_flush_batch. It might be better not to do the
extra
work of searching all pages in the extent in this case. This adds a my.cnf
variable
to allow that to be tested. The variable name is
"innodb_flush_neighbors_for_lru"
and the default is ON to match existing behavior.

Test Plan:
mtr, sysbench

Reviewed By: ryandmack