undo the fix for bug 49047 that was undone in official innodb
Summary:
There were two commits for 49047. We knew about the first but did not
know about the second. The second commit undoes some of the changes from the
first.
This change is the second commit. This also fixes a failure in
innodb_plugin.innodb_mysql
where lock-wait-timeout occurred instead of the expected deadlock.

Test Plan:
mtr

Reviewed By: ryandmack