RFC|DRAFT: InnoDB support for symlinked table files.
Summary:
With this change the following will happen:

  -If a .ibd file points to a non-existing path, then the table
corresponding to that .ibd file is created in that location.
  -After alter table add column, add index, delete column, rename table
operations the symlinks are still preserved.

Test Plan:
-Create symlink table.ibd that points to a non-existing table under the
data directory of a database.
  -Create table "table" using a mysql client. Make sure that MySQL uses
the symlinked file instead of creating a new one.
  -Alter table by adding/droping indexes/columns or renaming the table.
Make sure that the table still uses the symlinked file.

Reviewed By: ryandmack