Add a session variable that controls the block size for external merge-sort.
Summary:
This diff makes the block size that's used in the merge sort dynamic by adding
the session variable
innodb_merge_sort_block_size. The min value for this variable is 1M, and the
max is 1G. By varying the
values of this variable dbas can let the mysql server use larger blocks during
secondary index creation. The default
is 1048576 which is 1MB. See
https://www.intern.facebook.com/intern/wiki/index.php/Data/MySQL/MySQL_index_creation
for details on secondary index
creation.

Test Plan:
change the value of innodb_sort_block_size and notice the difference in time
spent for creating secondary indexes on
tables with 1M+ rows. Also add print statements that shows the block size and
the number of runs used in merge-sort.
Make sure that two different sessions can have different block sizes. Also make
sure that changing the block size
during merge-sort doesn't cause the operation to fail.

Reviewed By: mcallaghan
Reviewers: mcallaghan