Add log_query_sample_rate and log_error_sample_rate
Summary:
When log_query_sample_rate is > 0, then 1 / log_query_sample_rate of all
queries
are written to the slow query log. When log_error_sample_rate is > 0, then
1 / log_error_sample_rate of all queries that raise an error are written to the
slow query log.
These do not change behavior for the existing parameters that also determine
when
a statement is logged.
This also fixes the Errno value printed in the slow query log so that it is
usually correct.

Test Plan:
mtr

Reviewed By: ryandmack
Reviewers: ryandmack