Remove page compression log records caused by INSERTs.
Summary:
InnoDB logs the entire compressed page every time the page is
compressed/re-compressed in the redo log. These log
records contain the compressed data for the entire page image. Most of the time
this is unnecessary. This revision
fixes this partially by eliminating the compression log records caused by
INSERTs. Note that there will still be some
log records that contain the compressed page image because an INSERT can cause
the page to split, when that happens
the data on the page is split into two and placed on two different pages. After
this, the compressed images of both
of the pages are logged instead of the INSERT log records to both pages.

The main difference is in page_cur_insert_rec_zip_reorg() inside page0cur.c.
Previously this function called
page_zip_compress() with a non-NULL mtr to compress the page. When passed a
non-NULL mtr, page_zip_compress() logs the
image of the compressed page, therefore the INSERT log record is no longer
necessary. With this change,
page_zip_compress() is passed a NULL mtr and if it succeeds, the INSERT log
record and PAGE_ZIP_COMPRESS log record is
written.

We must rebuild xtrabackup in order to be able to recover from the transaction
logs produced by the new mysql binary that
contains this revision. This is because this revision adds a new log record
type and therefore it increases the max record type for the log
record types. This may cause assertions to fail in xtrabackup. Also the new log
records must be properly handled by xtrabackup.

Test Plan:
unit test: innodb_log_compressed_pages_crash.test performs a transaction that
inserts 21 records to a
compression-enabled table and then crashes the master at certain points during
execution. It makes sure that all 21
records remain on the table after recovery.

  51test innodb_log_compressed_pages_crash.test

stress test: innodb_master_replica_crash.test simulates multiple concurrent
transactions. It calls load_generator.py
which creates multiple threads each of which executes an OLTP-like load on the
database. The load_generator script
kills the database after a certain amount of time as specified by the .test
file. Then the .test script restarts the server and then
compares the checksums of the tables on both master and the replica to make
sure that the crash recovery succeeded.
This process is repeated multiple times.

  51test --big-test innodb_master_replica_crash.test

With the default parameters, the stress test can take up to 10 minutes. Change
the values in innodb_master_replica_crash.test if you want to make it shorter.
Also the entire state for all transactions is stored in the journal table. While
the
test is running, you can connect to the master server and run "SELECT * FROM
journal" to find out about the current state of the transactions. Since it takes
too much time, innodb_master_replica_crash is not included in the standard tests
and it
can only be run if the --big-test switch is turned on.

Reviewed By: mcallaghan
Reviewers: mcallaghan