Add crash recovery tests for rpl_transaction_enabled and catch one case where recovery cannot be done
Summary: This adds crash recovery tests for rpl_transaction_enabled with and with log-slave-updates
and fixes two bugs.
"
For background, when log-slave-updates is disabled on a slave then the sequence is:
* update innodb trx header page with slave offset
* innodb commit
* update relay-log.info
"
And when log-slave-updates is enabled the correct sequence is:
* update innodb trx header page with slave offset
* innodb prepare
* write XID to binlog and fsync it
* innodb commit
"
The first bug was not to update slave replication state in InnoDB on the prepare step when XA is
used on a slave (because log-slave-updates is set). If there is a crash after the XID event is
written to the binlog and before commit is done, during crash recovery the slave state in InnoDB
will be one commit behind the real state and the last transaction will be duplicated.
"
The second bug was to not detect a case where crash recovery is not possible.  When there is a
crash on the slave between the InnoDB prepare and logging the XID event to the slave binlog, then
on crash recovery rollback is done because the XID is not in the binlog. However there is no way
to rollback the slave replication state in the InnoDB system header. The workaround is to notice
the rollback and halt replication on startup with a message indicating that manual recovery might
be needed (replay the last transaction prior to the current slave offset as listed in InnoDB).
The error listed in SHOW SLAVE STATUS is ER_MASTER_INFO (1201).
"
The error in the database error log is:
  Rollback done for prepared transaction because its XID was not
  in the binlog. rpl_transaction_enabled does not handle this case.
  You might need to manually replay the last InnoDB transaction
  prior to the slave offset listed for InnoDB above.
"
The error above is only possible when log-slave-updates is used on the slave.
The fix for the error is to use global transaction IDs from the Google patch.
"
An example of the error is:
110226 20:01:48  InnoDB: Starting recovery for XA transactions...
110226 20:01:48  InnoDB: Transaction 302 in prepared state after recovery
110226 20:01:48  InnoDB: Transaction contains changes to 1 rows
110226 20:01:48  InnoDB: 1 transactions in prepared state after recovery
110226 20:01:48 [Note] Found 1 prepared transaction(s) in InnoDB
110226 20:01:48 [Note] rollback xid 'MySQLXid\2\0\0\0\0\0\0\0$\0\0\0\0\0\0\0'
110226 20:01:48 InnoDB: reading slave state from the transaction system header
                InnoDB: relay-log - filename ./slave-relay-bin.000003, position (849)
                InnoDB: master-log - filename master-bin.000001, position (659)
110226 20:01:48 [Note] Crash recovery finished.
110226 20:01:48 [Note] Read master.info: log_file_name: master-bin.000001, pos: 842
110226 20:01:48 [Note] init_relay_log_info: slave state from InnoDB: relay-log['./slave-relay-bin.000003'(849)], master-log['master-bin.000001'(659)]
110226 20:01:48 [ERROR] Slave SQL: Rollback done for prepared transaction because its XID was not in the binlog. rpl_transaction_enabled does not handle this case. You might need to manu
ally replay the last InnoDB transaction prior to the slave offset listed for InnoDB above. Error_code: 1201
110226 20:01:48 [ERROR] Rollback done for prepared transaction because its XID was not in the binlog. rpl_transaction_enabled does not handle this case. You might need to manually replay
 the last InnoDB transaction prior to the slave offset listed for InnoDB above.

Reviewers: chip

Test Plan: mtr