Remove compressed page image log records caused by UPDATEs and DELETEs.
Summary:
This diff removes the unnecessary compressed page image logging that was
performed because of UPDATEs. The DELETEs do not incur any compressed
page image log records, because they are performed by just marking
records as deleted. (Later a purge operation compactifies the pages that
have holes in them, but that too doesn't cause unnecessary page image
log records) UPDATEs are performed by first marking the existing record
as deleted, then inserting a new record therefore most of the
unnecessary page-image logging was already removed. The main difference
to innodb in this revision is in btr0cur.c, where a page is recompressed
to make room for an UPDATE operation.

Test Plan:
innodb_log_compressed_pages_crash_inserts.test : Tests the crash
recovery for an INSERT benchmark. This test inserts a number of records
and makes the innodb crash before and after commit. The records are
expected to be inserted if crash happened after the commit, and not
expected to be inserted if the crash happened before the commit.
innodb_log_compressed_pages_crash_deletes.test : Tests the crash
recovery for a DELETE benchmark. Similar to above. A table is emptied by
a number of DELETE queries. Table must be empty if the crash happens
after commit, it needs to have all of the records otherwise.
innodb_log_compressed_pages_crash_updates.test : Tests the crash
recovery for an UPDATE benchmark. Similar to above tests. All the rows
on a table are UPDATEd within a transaction. If the crash happens after
commit, then the the updated rows must be present after recovery,
otherwise the original records must be present.

Reviewed By: mcallaghan
Reviewers: mcallaghan