[Replace buf_block_buf_fix_inc  with bpage->buf_fix_count++]
Summary:
When porting the fix that percona gave, I replaced bpage->buf_fix_count++
with buf_block_buf_fix_inc as the latter does a sanity check on owning mutex
(block->mutex).
However, the buffer in question here holds compressed page and so buf pool zip
mutex and
not block->mutex should be held in this case. So that sanity check is not the
right one.
So replace buf_block_buf_fix_inc  with bpage->buf_fix_count++ as in the fix
that
percona gave.

Test Plan:
I'll work with Nizam to ensure the problem he is seeing is fixed by this.

Reviewed By: ryandmack
Reviewers: mcallaghan, ryandmack
Commenters: nizamordulu