Rewrite table_stats support in InnoDB to use per-tablespace counters
Summary:
This removes the use of tagging pages in the buffer pool with a
TABLE_STATS reference. That didn't capture all IO activity (such as reads
done by srv_master_thread for insert buffer merges and purge).
"
It is simpler to only report IO stats activity when file-per-table is
used and use the tablespace ID to assign the cost. To track this back
to a table, the db and table name for the tablespace must be saved,
so this change has somewhat messy code to unparse them from the
tablespace name. When file-per-table is not used, all tables will
be in the system tablespace and IO activity will only be reported
for it. The db name "sys:innodb" is used for the system tablespace,
logfiles and any tables for which a name cannot be unparsed. The
system tablespace has the name "system" in IS.table_statistics,
logfiles have the name "log" and tables that cannot be unparsed
have the name "other".
"
The stats are reported by InnoDB only when IS.table_statistics is
queried and the tablespace hash table is iterated to find entries
that have had IO activity since the last query. The cost of iteration
is linear in the hash table size.

Test Plan:
mtr, run in production and compare stats to old binary

Reviewed By: nizamordulu
Reviewers: nizamordulu, tuomaspelkonen
Commenters: tuomaspelkonen