Add my.cnf option allow_hint_to_missing_index
Summary:
This option is disabled by default. When set the use of an index name that
does not exist does not raise an error for FORCE INDEX, USE INDEX and IGNORE
INDEX.
When not set ER_KEY_DOES_NOT_EXITS is raised (not a typo, _EXITS is the
suffix).

Test Plan:
mtr

Reviewed By: dmituzas
Reviewers: dmituzas