Fix race with background drop table
Summary:
The problem is that if the thread that did a drop table using the background
drop table queue then calls row_drop_tables_for_mysql_in_background and
srv_master_thread
has called it concurrently, then srv_master_thread might have removed the table
from
the drop table list but not yet removed the table or the file for it. In that
case
the user thread that called the same function should wait for the pending drop
to
finish. I used sleep rather than pthread_cond_t to do the wait because
kernel_mutex
is not a pthread mutex and there are several other examples in InnoDB source
where
sleep is done like this (unfortunately).

Test Plan:
mtr, added extra calls to sleep to expose the race and show that this fixed it

Reviewed By: nizamordulu
Reviewers: nizamordulu
Commenters: chip