Fix mysql bug 61341 -- buf_LRU_insert_zip_clean is order N on LRU length
Summary:
This removes use of the zip_clean field from non-debug builds to avoid
performance stalls from O(N) operations. Insert into the list can be slow when
the LRU is mostly clean compressed pages, each compressed page has an
uncompress page on the unzip_LRU and InnoDB begins to evicted the uncompress
pages.
"
This is the fix for now per reading the code and feedback on
http://bugs.mysql.com/61341
from the InnoDB dev team. The problem is reproduced by sysbench.

Test Plan:
mtr

Reviewed By: chip
Reviewers: chip