Kludge for http://bugs.mysql.com/bug.php?id=60926, add slave_max_allowed_packet
Summary:
Add slave_max_allowed_packet so that packets between master and slave
can be larger than between client and master. This is required because the
master
has weak enforcement of max_allowed_packet and binlog events can be larger than
it.
It must be set on the master and slave. It is ignored when 0.

Test Plan: mtr
Reviewed By: chip
Reviewers: chip