Change transaction control to return errors rather than wait when limit exceeded
Summary: This changes transaction control to return an error when the limit is exceeded.
It also is based on the number of undo slots used per-account as a lack of undo slots
is what causes the too-many-transactions errors. Transactions can use 2 undo slots
in InnoDB -- one for insert and one for undo. This raises an error when an account
starts an INSERT/UPDATE/REPLACE/DELETE statement and the account is already using
more than max_concurrent_transactions. I repeat, this allows an account to use N+1
or N+2 undo slots when the limit is N. So it is a fuzzy limit, but that makes the
implementation simpler. The limit is also fuzzy because the check for entry is made
long before a session increments the count. So a few sessions can pass the check for
entry and then all increment the count. So be it, this allows a simple implementation
that I am willing to debug.
"
Updated perf results are at https://www.facebook.com/notes/mark-callaghan/max-concurrent-connections-queries-and-transactions/10150245619365933

Reviewers: chip

Test Plan: mtr, sysbench