Move admission control enforcement earlier
Summary:
This moves enforcement of admission control earlier in the call stack from
mysql_execute_command to dispatch_command. There are positive side-effects
1) not as many functions must pass a new argument so the diff is smaller
2) recursion isn't a problem
3) for multi-statement calls that Tao should use a connection only goes through
admission control once per RPC, not once per statement
The negative side-effect is that SHOW PROCESSLIST no longer shows the SQL for a
connection waiting on admission control. That is not visible because the query
has
yet to be parsed. Maybe I can fix that in the future.
"
But the real reason for this change is that commit for auto-commit statements
is
done after leaving mysql_parse. So when max_concurrenct_queries is 4 for an
account,
it can still have hundreds of connections dogpiling on commit. I noticed this
in
my sysbench tests and the previously shared results include this change.

Test Plan: mtr, sysbench perf results
Reviewed By: chip
Reviewers: chip